requirements.txt
# FastAPI and ASGI
fastapi>=0.110.0
uvicorn[standard]>=0.29.0

# Environment config
python-dotenv>=1.0.1
pydantic-settings>=2.2.1

# LangChain ecosystem
langchain>=0.1.14
langchain-core>=0.1.33
langchain-community>=0.0.30
langgraph>=0.0.40

# LLMs
langchain-groq>=0.3.6
langchain-google-genai>=0.0.8
google-generativeai>=0.4.1

# Vector DB
pinecone-client>=3.2.2
langchain-pinecone>=0.0.3

# Scraping and utilities
beautifulsoup4
httpx
aiofiles
tqdm

# HuggingFace + embeddings
transformers
sentence-transformers
huggingface-hub

# Optional utilities
tiktoken
regex
